requirements.txt
absl-py==1.3.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.3.1
asciitree==0.3.3
astor==0.8.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle==2.2.1
colorama==0.4.6
comet-ml==3.32.5
configobj==5.0.8
contextlib2==21.6.0
cycler==0.11.0
dill==0.3.5.1
dulwich==0.21.3
easydict==1.9
entrypoints==0.4
everett==3.1.0
fasteners==0.18
filelock==3.8.0
fonttools==4.37.3
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.9
GitPython==3.1.26
glib==1.0.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
got==0.0.1
grpcio==1.49.1
higher==0.2.1
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.12.0
JIT==0.0.1
joblib==1.2.0
json-tricks==3.16.1
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
kiwisolver==1.4.4
kornia==0.6.7
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
multidict==6.0.2
multiprocess==0.70.13
networkx==2.6.3
nni==2.10
numba==0.56.3
numcodecs==0.10.2
numpy==1.21.6
oauthlib==3.2.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
pkgconfig==1.5.5
pkgutil_resolve_name==1.3.10
prettytable==3.6.0
protobuf==3.19.6
psutil==5.9.4
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
pyrsistent==0.19.3
python-box==6.1.0
python-dateutil==2.8.2
PythonWebHDFS==0.2.3
pytz==2022.2.1

PyYAML==6.0
randaugment==1.0.2
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
responses==0.18.0
rsa==4.9
schema==0.7.5
scikit-learn==1.0.2
scipy==1.7.3
semantic-version==2.10.0
sentry-sdk==1.18.0
setproctitle==1.3.2
shapely==2.0.1
simplejson==3.18.3
six==1.16.0
smmap==5.0.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.13
torch==1.12.0
torchnet==0.0.4
torchvision==0.13.0
tornado==6.2
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.3.0
urllib3==1.26.12
visdom==0.2.3
wcwidth==0.2.6
websocket-client==1.3.3
websockets==10.4
Werkzeug==2.2.2
wrapt==1.15.0
wurlitzer==3.0.3
xxhash==3.1.0
yarl==1.8.1
zarr==2.12.0
zipp==3.9.0
pyvips